+++ lib/src/chinese_helper.dart
@@ -1,5 +1,6 @@
 import 'package:lpinyin/src/pinyin_resource.dart';
 
+/// Chinese Helper.
 class ChineseHelper {
   static final String chineseRegex = "[\\u4e00-\\u9fa5]";
   static final RegExp chineseRegexp = new RegExp(chineseRegex);