+++ lib/src/pinyin_helper.dart
@@ -25,13 +25,12 @@ class PinyinHelper {
   /**
    *获取字符串首字拼音
    *@param str 需要转换的字符串
-   *  @return 首字拼音
+   *  @return 首字拼音 (成都 cheng)
    */
 
   ///
   static String getFirstWordPinyin(String str) {
-    String _pinyin =
-        convertToPinyinStringWithoutException(str, separator: pinyinSeparator);
+    String _pinyin = getPinyinE(str, separator: pinyinSeparator);
     return _pinyin.split(pinyinSeparator)[0];
   }
 
@@ -39,14 +38,14 @@ class PinyinHelper {
    * 获取字符串对应拼音的首字母
    * @param str 需要转换的字符串
    * @param defPinyin 拼音分隔符 def: '#'
-   * @return 对应拼音的首字母
+   * @return 对应拼音的首字母 (成都 cd)
    */
 
   ///
   static String getShortPinyin(String str, {String defPinyin: '#'}) {
     StringBuffer sb = new StringBuffer();
-    String pinyin = convertToPinyinStringWithoutException(str,
-        separator: pinyinSeparator, defPinyin: defPinyin);
+    String pinyin =
+        getPinyinE(str, separator: pinyinSeparator, defPinyin: defPinyin);
     List<String> list = pinyin.split(pinyinSeparator);
     list.forEach((value) {
       sb.write(value[0]);
@@ -59,11 +58,11 @@ class PinyinHelper {
    * @param str 需要转换的字符串
    * @param separator 拼音分隔符 def: " "
    * @param format 拼音格式 def: PinyinFormat.WITHOUT_TONE
-   * @return 字符串的拼音
+   * @return 字符串的拼音(成都 cheng du)
    */
 
   ///
-  static String convertToPinyinString(String str,
+  static String getPinyin(String str,
       {String separator: " ", PinyinFormat format: PinyinFormat.WITHOUT_TONE}) {
     StringBuffer sb = new StringBuffer();
     str = ChineseHelper.convertToSimplifiedChinese(str);
@@ -99,18 +98,18 @@ class PinyinHelper {
   }
 
   /**
-   * 将字符串转换成相应格式的拼音 (不能转换的字拼音默认用'#'替代 )
+   * 将字符串转换成相应格式的拼音 (不能转换的字拼音默认用' '替代 )
    * @param str 需要转换的字符串
    * @param separator 拼音分隔符 def: " "
-   * @param defPinyin 拼音分隔符 def: '#'
+   * @param defPinyin 拼音分隔符 def: ' '
    * @param format 拼音格式 def: PinyinFormat.WITHOUT_TONE
-   * @return 字符串的拼音
+   * @return 字符串的拼音(成都 cheng du)
    */
 
   ///
-  static String convertToPinyinStringWithoutException(String str,
+  static String getPinyinE(String str,
       {String separator: " ",
-      String defPinyin: '#',
+      String defPinyin: ' ',
       PinyinFormat format: PinyinFormat.WITHOUT_TONE}) {
     StringBuffer sb = new StringBuffer();
     str = ChineseHelper.convertToSimplifiedChinese(str);
@@ -314,17 +313,18 @@ class PinyinHelper {
     return false;
   }
 
-  ///添加拼音字典
+  /// 添加拼音字典
   static void addPinyinDict(List<String> list) {
     pinyinMap.addAll(PinyinResource.getResource(list));
   }
 
-  ///添加多音字字典
+  /// 添加多音字字典
   static void addMultiPinyinDict(List<String> list) {
     multiPinyinMap.addAll(PinyinResource.getResource(list));
   }
 }
 
+/// 多音字
 class MultiPinyin {
   String word;
   String pinyin;
@@ -333,13 +333,10 @@ class MultiPinyin {
 
   @override
   String toString() {
-    return " {" +
-        " \"word\":\"" +
-        word +
-        "\"" +
-        ", \"pinyin\":\"" +
-        pinyin +
-        "\"" +
-        '}';
+    StringBuffer sb = new StringBuffer('{');
+    sb.write("\"word\":\"$word\"");
+    sb.write(",\"pinyin\":\"$pinyin\"");
+    sb.write('}');
+    return sb.toString();
   }
 }